+++ CHANGELOG.md
@@ -19,6 +19,7 @@ It contains the time at which a pre-detection was created by the processor.
 * reimplemented kafka input connector
   - move kafka config options to `kafka_config` dictionary
   - implement manual commit behaviour if `enable.auto.commit: false`
+  - implement on_commit callback to check for errors during commit
 * reimplement kafka output connector
   - move kafka config options to `kafka_config` dictionary